dcaw-backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beanie==2.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
lazy-model==0.3.0
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1